refactor algorithm package to live under /parsec
