+++ convert_utils.go
@@ -7,10 +7,10 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/parallaxsecond/parsec-client-go/algorithm"
 	"github.com/parallaxsecond/parsec-client-go/interface/auth"
 	"github.com/parallaxsecond/parsec-client-go/interface/operations/listauthenticators"
 	"github.com/parallaxsecond/parsec-client-go/interface/operations/psaalgorithm"
+	"github.com/parallaxsecond/parsec-client-go/parsec/algorithm"
 )
 
 func newAuthenticatorInfoFromOp(inf *listauthenticators.AuthenticatorInfo) (*auth.AuthenticatorInfo, error) {
