+++ e2etest/signing_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	parsec "github.com/parallaxsecond/parsec-client-go"
-	"github.com/parallaxsecond/parsec-client-go/algorithm"
+	"github.com/parallaxsecond/parsec-client-go/parsec/algorithm"
 )
 
 func TestSigning(t *testing.T) {
